Address review comments: move new_poller() to internal too,
