+++ generator/internal/rust/templates/crate/src/builder.rs.mustache
@@ -169,7 +169,7 @@ pub mod {{Codec.ModuleName}} {
                 Ok(Operation::new(op))
             };
 
-            lro::new_poller(polling_error_policy, polling_backoff_policy, start, query)
+            lro::internal::new_poller(polling_error_policy, polling_backoff_policy, start, query)
         }
         {{/OperationInfo}}
         {{#InputType.Codec.BasicFields}}
