+++ src/lro/tests/fake/library/builders.rs
@@ -90,7 +90,12 @@ impl CreateResource {
             let op = self.send().await?;
             Ok(Operation::new(op))
         };
-        google_cloud_lro::new_poller(polling_error_policy, polling_backoff_policy, start, query)
+        google_cloud_lro::internal::new_poller(
+            polling_error_policy,
+            polling_backoff_policy,
+            start,
+            query,
+        )
     }
 }
